move usage stat bundle implementation to usage stats
