+++ pkg/infra/usagestats/service/usage_stats_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/grafana/grafana/pkg/infra/usagestats"
 	"github.com/grafana/grafana/pkg/plugins"
 	"github.com/grafana/grafana/pkg/services/accesscontrol/actest"
+	"github.com/grafana/grafana/pkg/services/supportbundles/supportbundlestest"
 	"github.com/grafana/grafana/pkg/setting"
 )
 
@@ -223,6 +224,7 @@ func createService(t *testing.T, cfg setting.Cfg, sqlStore db.DB, withDB bool) *
 		tracing.InitializeTracerForTest(),
 		actest.FakeAccessControl{ExpectedDisabled: true},
 		actest.FakeService{},
+		supportbundlestest.NewFakeBundleService(),
 	)
 
 	return service
